requirements-dev.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-config-url==0.1.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
djangorestframework==3.12.2
idna==2.10
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.4
